addressed review comments SimulatorStrategyMultiNodeSingleSim
