--- planner/scheduler/testdata/prefer-large-nodes-config.yaml
@@ -1,23 +1,0 @@
-apiVersion: kubescheduler.config.k8s.io/v1
-kind: KubeSchedulerConfiguration
-clientConnection:
-  qps: 100
-  burst: 50
-  contentType: application/json
-  kubeconfig: /tmp/minkapi.yaml
-enableProfiling: true
-leaderElection:
-  leaderElect: false
-profiles:
-  - schedulerName: default-scheduler
-  - schedulerName: bin-packing-scheduler
-    pluginConfig:
-      - name: NodeResourcesFit
-        args:
-          scoringStrategy:
-            type: MostAllocated
-    plugins:
-      score:
-        disabled:
-          - name: NodeResourcesBalancedAllocation
-
