+++ planner/simulator/scaleout/multinode/multinode.go
@@ -3,7 +3,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 // Package multinode provides implementation and helper routines of a ScaleOutSimulator that performs simulations that scale
-// multiple nodes at a time
+// multiple nodes for a single scale-out simulation
 package multinode
 
 import (
@@ -13,8 +13,10 @@ import (
 	"github.com/gardener/scaling-advisor/planner/simulator/scaleout"
 
 	commontypes "github.com/gardener/scaling-advisor/api/common/types"
+	sacorev1alpha1 "github.com/gardener/scaling-advisor/api/core/v1alpha1"
 	"github.com/gardener/scaling-advisor/api/minkapi"
 	plannerapi "github.com/gardener/scaling-advisor/api/planner"
+	"github.com/gardener/scaling-advisor/common/objutil"
 	"github.com/go-logr/logr"
 )
 
@@ -22,7 +24,7 @@ var (
 	_ plannerapi.ScaleOutSimulator = (*simulatorSingleSim)(nil)
 )
 
-// simulatorSingleSim is a Simulator that implements ScaleOutSimulator for the SimulatorStrategyMultiNodeSingleSim.
+// simulatorSingleSim is a Simulator that implements [plannerapi.ScaleOutSimulator] for the [commontypes.SimulatorStrategyMultiNodeSingleSim].
 type simulatorSingleSim struct {
 	viewAccess        minkapi.ViewAccess
 	schedulerLauncher plannerapi.SchedulerLauncher
@@ -32,7 +34,7 @@ type simulatorSingleSim struct {
 	simulatorConfig   plannerapi.SimulatorConfig
 }
 
-// New creates a new plannerapi.ScaleOutSimulator that runs simulations sequentially scaling multiple nodes from
+// New creates a new [plannerapi.ScaleOutSimulator] that runs simulations sequentially scaling multiple nodes from
 // different NodeTemplates at the same priority.
 func New(args plannerapi.SimulatorArgs) (plannerapi.ScaleOutSimulator, error) {
 	return &simulatorSingleSim{
@@ -71,51 +73,6 @@ func (s *simulatorSingleSim) doSimulate(ctx context.Context) (err error) {
 	return
 }
 
-func (s *simulatorSingleSim) runAllGroups(ctx context.Context) (err error) {
-	var (
-		log           = logr.FromContextOrDiscard(ctx)
-		groupPassView = s.state.RequestView()
-		simResults    []plannerapi.ScaleOutSimResult
-		allSimResults []plannerapi.ScaleOutSimResult
-	)
-	for groupIndex, group := range s.state.SimulationGroups {
-		log := log.WithValues("groupIndex", groupIndex, "groupName", group.Name()) // in-loop log enhanced with further params
-		passCtx := logr.NewContext(ctx, log)
-		if simResults, groupPassView, err = s.runPassForGroup(passCtx, group, groupPassView); err != nil {
-			return
-		}
-		if len(simResults) > 0 {
-			allSimResults = append(allSimResults, simResults...)
-		}
-	}
-	if s.state.Request.AdviceGenerationMode.IsAllAtOnce() {
-		err = scaleout.SendPlanResultUsingSimResults(ctx, s.state.ResultCh, s.state.Request, s.state.SimRunCounter.Load(), allSimResults)
-	}
-	return
-}
-
-func (s *simulatorSingleSim) runPassForGroup(ctx context.Context, group plannerapi.ScaleOutSimGroup, groupPassView minkapi.View) (simResults []plannerapi.ScaleOutSimResult, nextGroupPassView minkapi.View, err error) {
-	//var (
-	//	log = logr.FromContextOrDiscard(ctx)
-	//)
-	simResults, err = group.Run(ctx, func(ctx context.Context, name string) (minkapi.View, error) {
-		return s.state.CreateSandboxView(ctx, name, groupPassView)
-	})
-	if err != nil {
-		return
-	}
-	if len(simResults) == 0 {
-		nextGroupPassView = groupPassView
-		return
-	} else {
-		nextGroupPassView = simResults[0].View // all simResults share the same View in this strategy
-	}
-	if s.state.Request.AdviceGenerationMode.IsIncremental() {
-		err = scaleout.SendPlanResultUsingSimResults(ctx, s.state.ResultCh, s.state.Request, s.state.SimRunCounter.Load(), simResults)
-	}
-	return
-}
-
 func (s *simulatorSingleSim) createAndGroupSimulations(ctx context.Context) ([]plannerapi.ScaleOutSimGroup, error) {
 	var (
 		allScaleOutNodeTemplates = scaleout.CreateAllNodeTemplates(s.state.Request.Constraint.Spec.NodePools)
@@ -144,4 +101,76 @@ func (s *simulatorSingleSim) createAndGroupSimulations(ctx context.Context) ([]p
 		simNum++
 	}
 	return scaleout.CreateScaleOutSimGroups(s.state.Request.GetRef(), allSimulations)
+}
+
+func (s *simulatorSingleSim) runAllGroups(ctx context.Context) (err error) {
+	var (
+		log           = logr.FromContextOrDiscard(ctx)
+		groupPassView = s.state.RequestView()
+		simResults    []plannerapi.ScaleOutSimResult
+		allSimResults []plannerapi.ScaleOutSimResult
+	)
+	for groupIndex, group := range s.state.SimulationGroups {
+		log := log.WithValues("groupIndex", groupIndex, "groupName", group.Name()) // in-loop log enhanced with further params
+		passCtx := logr.NewContext(ctx, log)
+		if simResults, groupPassView, err = s.runPassForGroup(passCtx, group, groupPassView); err != nil {
+			return
+		}
+		if len(simResults) > 0 {
+			allSimResults = append(allSimResults, simResults...)
+		}
+	}
+	if s.state.Request.AdviceGenerationMode.IsAllAtOnce() {
+		err = sendPlanResultUsingSimResults(ctx, s.state.ResultCh, s.state.Request, s.state.SimRunCounter.Load(), allSimResults)
+	}
+	return
+}
+
+func (s *simulatorSingleSim) runPassForGroup(ctx context.Context, group plannerapi.ScaleOutSimGroup, groupPassView minkapi.View) (simResults []plannerapi.ScaleOutSimResult, nextGroupPassView minkapi.View, err error) {
+	simResults, err = group.Run(ctx, func(ctx context.Context, name string) (minkapi.View, error) {
+		return s.state.CreateSandboxView(ctx, name, groupPassView)
+	})
+	if err != nil {
+		return
+	}
+	if len(simResults) == 0 {
+		nextGroupPassView = groupPassView
+		return
+	}
+	nextGroupPassView = simResults[0].View // all simResults share the same View in this strategy
+	if s.state.Request.AdviceGenerationMode.IsIncremental() {
+		err = sendPlanResultUsingSimResults(ctx, s.state.ResultCh, s.state.Request, s.state.SimRunCounter.Load(), simResults)
+	}
+	return
+}
+
+// sendPlanResultUsingSimResults constraints a [plannerapi.ScaleOutPlanResult] from the given slice of
+// [plannerapi.ScaleOutSimResult] and referring the given [plannerapi.Request] and sends the same on the given result
+// channel.
+func sendPlanResultUsingSimResults(ctx context.Context,
+	resultCh chan<- plannerapi.ScaleOutPlanResult,
+	req *plannerapi.Request, simulationRunCount uint32, // TODO: introduce a plannerapi.Metrics.
+	simResults []plannerapi.ScaleOutSimResult) error {
+	log := logr.FromContextOrDiscard(ctx)
+	labels := scaleout.CreatePlanLabels(req, simulationRunCount)
+	existingNodeCountByPlacement, err := req.Snapshot.GetNodeCountByPlacement()
+	if err != nil {
+		return err
+	}
+	var scaleOutPlan sacorev1alpha1.ScaleOutPlan
+	for _, sr := range simResults {
+		for _, item := range sr.Items {
+			existingCount := existingNodeCountByPlacement[item.NodePlacement]
+			item.CurrentReplicas = existingCount
+			scaleOutPlan.Items = append(scaleOutPlan.Items, item)
+		}
+		scaleOutPlan.UnsatisfiedPodNames = objutil.GetFullNames(sr.LeftoverUnscheduledPods)
+	}
+	planResult := plannerapi.ScaleOutPlanResult{
+		Labels:       labels,
+		ScaleOutPlan: &scaleOutPlan,
+	}
+	log.V(2).Info("Sent Planner Success Response", "response", planResult)
+	resultCh <- planResult
+	return nil
 }
